Split the DMPlex boundary tests into more files.
